pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devx-ai"
version = "0.1.0"
description = "AI-powered developer tooling platform — code review, test generation, SDLC automation, and engineering metrics."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "devx-ai contributors" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.27,<1.0",
    "httpx>=0.27,<1.0",
    "openai>=1.12,<2.0",
    "anthropic>=0.21,<1.0",
    "tree-sitter>=0.22,<1.0",
    "tree-sitter-python>=0.22,<1.0",
    "tree-sitter-javascript>=0.22,<1.0",
    "pyyaml>=6.0,<7.0",
    "rich>=13.0,<14.0",
    "tenacity>=8.2,<10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<6.0",
    "pytest-asyncio>=0.23,<1.0",
    "mypy>=1.8,<2.0",
    "ruff>=0.3,<1.0",
    "pre-commit>=3.6,<4.0",
    "respx>=0.21,<1.0",
]

[project.urls]
Homepage = "https://github.com/cortexark/devx-ai"
Documentation = "https://github.com/cortexark/devx-ai/tree/main/docs"
Repository = "https://github.com/cortexark/devx-ai"
Issues = "https://github.com/cortexark/devx-ai/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/devx"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers --tb=short -q"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external services",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["tree_sitter.*", "tree_sitter_python.*", "tree_sitter_javascript.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = [
    "SIM102",  # collapsible-if: sometimes nested ifs are clearer
]

[tool.ruff.lint.isort]
known-first-party = ["devx"]

[tool.coverage.run]
source = ["src/devx"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 75
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
]
